requirements.txt
aiogram
sqlalchemy
asyncpg
greenlet
pydantic-settings
AiogramStorages